+++ src/centrality/eigenvector.c
@@ -28,7 +28,9 @@
 
 #include "centrality/centrality_internal.h"
 
+#include <float.h>
 #include <limits.h>
+#include <math.h>
 
 /* Multiplies vector 'from' by the unweighted adjacency matrix and stores the result in 'to'. */
 static igraph_error_t adjmat_mul_unweighted(igraph_real_t *to, const igraph_real_t *from,
@@ -82,8 +84,32 @@ static igraph_error_t adjmat_mul_weighted(igraph_real_t *to, const igraph_real_t
     return IGRAPH_SUCCESS;
 }
 
+/* Checks if any eigenvector centrality values are zero. If they are, it indicates that the
+ * graph is not (strongly) connected. Eigenvector centrality is not meaningful for such graphs.
+ * To account for numerical inaccuracies, a threshold of 'eps' is used when testing for zero.
+ * This function is intended to be used with eigenvector centrality values scaled such that
+ * the maximum is 1. 'eps' is chosen accordinly.
+ */
+static void warn_zero_entries(const igraph_vector_t *evcent) {
+    /* This is a conservative tolerance that will still catch most values
+     * which should be zero without being triggered by small yet truly
+     * nonzero values.
+     * See https://github.com/igraph/igraph/pull/2592 */
+    const igraph_real_t tol = 10 * DBL_EPSILON;
+    const igraph_integer_t n = igraph_vector_size(evcent);
+
+    for (igraph_integer_t i=0; i < n; i++) {
+        igraph_real_t x = VECTOR(*evcent)[i];
+        if (-tol < x && x < tol) {
+            IGRAPH_WARNING("Some eigenvector centralities are nearly zero, indicating that the graph may not be (strongly) connected. "
+                           "Eigenvector centrality is not meaningful for disconnected graphs.");
+            return;
+        }
+    }
+}
+
 static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t *graph, igraph_vector_t *vector,
-                                                      igraph_real_t *value, igraph_bool_t scale,
+                                                      igraph_real_t *value,
                                                       const igraph_vector_t *weights,
                                                       igraph_arpack_options_t *options) {
 
@@ -92,13 +118,14 @@ static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t 
     igraph_vector_t degree;
     igraph_integer_t i;
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
+    igraph_integer_t no_of_edges = igraph_ecount(graph);
     igraph_bool_t negative_weights = false;
 
     if (no_of_nodes > INT_MAX) {
         IGRAPH_ERROR("Graph has too many vertices for ARPACK.", IGRAPH_EOVERFLOW);
     }
 
-    if (igraph_ecount(graph) == 0) {
+    if (no_of_edges == 0) {
         /* special case: empty graph */
         if (value) {
             *value = 0;
@@ -107,16 +134,20 @@ static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t 
             IGRAPH_CHECK(igraph_vector_resize(vector, igraph_vcount(graph)));
             igraph_vector_fill(vector, 1);
         }
+        if (no_of_nodes > 1) {
+            IGRAPH_WARNING("The graph has no edges and is disconnected. "
+                           "Eigenvector centrality is not meaningful for disconnected graphs.");
+        }
         return IGRAPH_SUCCESS;
     }
 
     if (weights) {
         igraph_real_t min, max;
 
-        if (igraph_vector_size(weights) != igraph_ecount(graph)) {
+        if (igraph_vector_size(weights) != no_of_edges) {
             IGRAPH_ERRORF("Weights vector length (%" IGRAPH_PRId ") not equal to "
                     "number of edges (%" IGRAPH_PRId ").", IGRAPH_EINVAL,
-                    igraph_vector_size(weights), igraph_ecount(graph));
+                    igraph_vector_size(weights), no_of_edges);
         }
         /* Safe to call minmax, ecount == 0 case was caught earlier */
         igraph_vector_minmax(weights, &min, &max);
@@ -128,6 +159,10 @@ static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t 
             if (vector) {
                 IGRAPH_CHECK(igraph_vector_resize(vector, igraph_vcount(graph)));
                 igraph_vector_fill(vector, 1);
+            }
+            if (no_of_nodes > 1) {
+                IGRAPH_WARNING("All edge weights are zero, making the graph effectively disconnected. "
+                               "Eigenvector centrality is not meaningful for disconnected graphs.");
             }
             return IGRAPH_SUCCESS;
         }
@@ -202,37 +237,36 @@ static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t 
         igraph_integer_t which = 0;
         IGRAPH_CHECK(igraph_vector_resize(vector, no_of_nodes));
 
-        if (!negative_weights && VECTOR(values)[0] <= 0) {
-            /* Pathological case: largest eigenvalue is zero, therefore all the
-             * scores can also be zeros, this will be a valid eigenvector.
-             * This usually happens with graphs that have lots of sinks and
-             * sources only. */
-            igraph_vector_fill(vector, 0);
-            VECTOR(values)[0] = 0;
-        } else {
-            for (i = 0; i < no_of_nodes; i++) {
-                igraph_real_t tmp;
-                VECTOR(*vector)[i] = MATRIX(vectors, i, 0);
-                tmp = fabs(VECTOR(*vector)[i]);
-                if (tmp > amax) {
-                    amax = tmp;
-                    which = i;
-                }
+        /* Note: With non-negative weights, a zero eigenvalue should only occur
+         * when there are no edges, or all edge have zero weight. This case is
+         * caught earlier. Thus we do not handle the case of zero eigenvalues here .*/
+
+        for (i = 0; i < no_of_nodes; i++) {
+            igraph_real_t tmp;
+            VECTOR(*vector)[i] = MATRIX(vectors, i, 0);
+            tmp = fabs(VECTOR(*vector)[i]);
+            if (tmp > amax) {
+                amax = tmp;
+                which = i;
             }
-            if (scale && amax != 0) {
-                igraph_vector_scale(vector, 1 / VECTOR(*vector)[which]);
-            } else if (igraph_i_vector_mostly_negative(vector)) {
-                igraph_vector_scale(vector, -1.0);
+        }
+
+        /* Scale result so that the largest value is 1.0. */
+        if (amax != 0) {
+            igraph_vector_scale(vector, 1 / VECTOR(*vector)[which]);
+        } else if (igraph_i_vector_mostly_negative(vector)) {
+            igraph_vector_scale(vector, -1.0);
+        }
+
+        /* Correction for numeric inaccuracies (eliminating -0.0) */
+        if (! negative_weights) {
+            for (i = 0; i < no_of_nodes; i++) {
+                if (VECTOR(*vector)[i] < 0) {
+                    VECTOR(*vector)[i] = 0;
+                }
             }
 
-            /* Correction for numeric inaccuracies (eliminating -0.0) */
-            if (! negative_weights) {
-                for (i = 0; i < no_of_nodes; i++) {
-                    if (VECTOR(*vector)[i] < 0) {
-                        VECTOR(*vector)[i] = 0;
-                    }
-                }
-            }
+            warn_zero_entries(vector);
         }
     }
 
@@ -252,7 +286,8 @@ static igraph_error_t igraph_i_eigenvector_centrality_undirected(const igraph_t 
 }
 
 static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *graph, igraph_vector_t *vector,
-                                                    igraph_real_t *value, igraph_bool_t scale,
+                                                    igraph_real_t *value,
+                                                    igraph_neimode_t mode,
                                                     const igraph_vector_t *weights,
                                                     igraph_arpack_options_t *options) {
 
@@ -261,10 +296,12 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
     igraph_vector_t indegree;
     igraph_bool_t dag;
     igraph_integer_t no_of_nodes = igraph_vcount(graph);
-    igraph_integer_t i;
+    igraph_integer_t no_of_edges = igraph_ecount(graph);
     igraph_bool_t negative_weights = false;
 
-    if (igraph_ecount(graph) == 0) {
+    mode = IGRAPH_REVERSE_MODE(mode);
+
+    if (no_of_edges == 0) {
         /* special case: empty graph */
         if (value) {
             *value = 0;
@@ -272,6 +309,10 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
         if (vector) {
             IGRAPH_CHECK(igraph_vector_resize(vector, igraph_vcount(graph)));
             igraph_vector_fill(vector, 1);
+        }
+        if (no_of_nodes > 1) {
+            IGRAPH_WARNING("The graph has no edges and is disconnected. "
+                           "Eigenvector centrality is not meaningful for disconnected graphs.");
         }
         return IGRAPH_SUCCESS;
     }
@@ -281,7 +322,6 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
     IGRAPH_CHECK(igraph_is_dag(graph, &dag));
     if (dag) {
         /* special case: graph is a DAG */
-        IGRAPH_WARNING("Graph is directed and acyclic; eigenvector centralities will be zeros.");
         if (value) {
             *value = 0;
         }
@@ -289,16 +329,19 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
             IGRAPH_CHECK(igraph_vector_resize(vector, igraph_vcount(graph)));
             igraph_vector_fill(vector, 0);
         }
+        IGRAPH_WARNING("Graph is directed and acyclic, and therefore not strongly connected. "
+                       "Eigenvector centrality is not meaningful for disconnected graphs. "
+                       "All-zero centrality scores will be returned.");
         return IGRAPH_SUCCESS;
     }
 
     if (weights) {
         igraph_real_t min, max;
 
-        if (igraph_vector_size(weights) != igraph_ecount(graph)) {
+        if (igraph_vector_size(weights) != no_of_edges) {
             IGRAPH_ERRORF("Weights vector length (%" IGRAPH_PRId ") not equal to "
                     "number of edges (%" IGRAPH_PRId ").", IGRAPH_EINVAL,
-                    igraph_vector_size(weights), igraph_ecount(graph));
+                    igraph_vector_size(weights), no_of_edges);
         }
 
         /* Safe to call minmax, ecount == 0 case was caught earlier */
@@ -318,6 +361,10 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
             if (vector) {
                 IGRAPH_CHECK(igraph_vector_resize(vector, igraph_vcount(graph)));
                 igraph_vector_fill(vector, 1);
+            }
+            if (no_of_nodes > 1) {
+                IGRAPH_WARNING("All edge weights are zero, making the graph effectively disconnected. "
+                               "Eigenvector centrality is not meaningful for disconnected graphs.");
             }
             return IGRAPH_SUCCESS;
         }
@@ -340,9 +387,9 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
 
     IGRAPH_VECTOR_INIT_FINALLY(&indegree, no_of_nodes);
     IGRAPH_CHECK(igraph_strength(graph, &indegree, igraph_vss_all(),
-                                 IGRAPH_IN, IGRAPH_LOOPS, weights));
+                                 mode, IGRAPH_LOOPS, weights));
     RNG_BEGIN();
-    for (i = 0; i < no_of_nodes; i++) {
+    for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
         if (VECTOR(indegree)[i]) {
             MATRIX(vectors, i, 0) = VECTOR(indegree)[i] + RNG_UNIF(-1e-4, 1e-4);
         } else {
@@ -356,7 +403,7 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
     if (!weights) {
         igraph_adjlist_t adjlist;
 
-        IGRAPH_CHECK(igraph_adjlist_init(graph, &adjlist, IGRAPH_IN, IGRAPH_LOOPS_ONCE, IGRAPH_MULTIPLE));
+        IGRAPH_CHECK(igraph_adjlist_init(graph, &adjlist, mode, IGRAPH_LOOPS_ONCE, IGRAPH_MULTIPLE));
         IGRAPH_FINALLY(igraph_adjlist_destroy, &adjlist);
 
         IGRAPH_CHECK(igraph_arpack_rnsolve(adjmat_mul_unweighted,
@@ -373,7 +420,7 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
         data.inclist = &inclist;
         data.weights = weights;
 
-        IGRAPH_CHECK(igraph_inclist_init(graph, &inclist, IGRAPH_IN, IGRAPH_LOOPS_ONCE));
+        IGRAPH_CHECK(igraph_inclist_init(graph, &inclist, mode, IGRAPH_LOOPS_ONCE));
         IGRAPH_FINALLY(igraph_inclist_destroy, &inclist);
 
         IGRAPH_CHECK(igraph_arpack_rnsolve(adjmat_mul_weighted,
@@ -397,7 +444,7 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
             igraph_vector_fill(vector, 0);
             MATRIX(values, 0, 0) = 0;
         } else {
-            for (i = 0; i < no_of_nodes; i++) {
+            for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
                 igraph_real_t tmp;
                 VECTOR(*vector)[i] = MATRIX(vectors, i, 0);
                 tmp = fabs(VECTOR(*vector)[i]);
@@ -406,7 +453,9 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
                     which = i;
                 }
             }
-            if (scale && amax != 0) {
+
+            /* Scale result so that the largest value is 1.0. */
+            if (amax != 0) {
                 igraph_vector_scale(vector, 1 / VECTOR(*vector)[which]);
             } else if (igraph_i_vector_mostly_negative(vector)) {
                 igraph_vector_scale(vector, -1.0);
@@ -415,11 +464,13 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
 
         /* Correction for numeric inaccuracies (eliminating -0.0) */
         if (! negative_weights) {
-            for (i = 0; i < no_of_nodes; i++) {
+            for (igraph_integer_t i = 0; i < no_of_nodes; i++) {
                 if (VECTOR(*vector)[i] < 0) {
                     VECTOR(*vector)[i] = 0;
                 }
             }
+
+            warn_zero_entries(vector);
         }
     }
 
@@ -462,14 +513,8 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
  * are added up.
  *
  * </para><para>
- * The centrality scores returned by igraph can be normalized
- * (using the \p scale parameter) such that the largest eigenvector centrality
- * score is 1 (with one exception, see below).
- *
- * </para><para>
- * In the directed case, the left eigenvector of the adjacency matrix is
- * calculated. In other words, the centrality of a vertex is proportional
- * to the sum of centralities of vertices pointing to it.
+ * The centrality scores returned by igraph are normalized such that the largest
+ * eigenvector centrality score is 1, unless all scores are zeros.
  *
  * </para><para>
  * Eigenvector centrality is meaningful only for (strongly) connected graphs.
@@ -499,10 +544,22 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
  *     be a null pointer, then it is ignored.
  * \param value If not a null pointer, then the eigenvalue
  *     corresponding to the found eigenvector is stored here.
- * \param directed Boolean scalar, whether to consider edge directions
- *     in a directed graph. It is ignored for undirected graphs.
- * \param scale If not zero then the result will be scaled such that
- *     the absolute value of the maximum centrality is one.
+ * \param mode How to consider edge directions in directed graphs.
+ *     It is ignored for undirected graphs. Possible values:
+ *        \clist
+ *        \cli IGRAPH_OUT
+ *          the left eigenvector of the adjacency matrix is calculated,
+ *          i.e. the centrality of a vertex is proportional to the sum
+ *          of centralities of vertices pointing to it. This is the standard
+ *          eigenvector centrality.
+ *        \cli IGRAPH_IN
+ *          the right eigenvector of the adjacency matrix is calculated,
+ *          i.e. the centrality of a vertex is proportional to the sum
+ *          of centralities of vertices it points to.
+ *        \cli IGRAPH_ALL
+ *          edge directions are ignored, and the unweighted eigenvector
+ *          centrality is calculated.
+ *        \endclist
  * \param weights A null pointer (indicating no edge weights), or a vector
  *     giving the weights of the edges. Weights should be positive to guarantee
  *     a meaningful result. The algorithm might produce complex numbers when some
@@ -528,7 +585,7 @@ static igraph_error_t igraph_i_eigenvector_centrality_directed(const igraph_t *g
 igraph_error_t igraph_eigenvector_centrality(const igraph_t *graph,
                                   igraph_vector_t *vector,
                                   igraph_real_t *value,
-                                  igraph_bool_t directed, igraph_bool_t scale,
+                                  igraph_neimode_t mode,
                                   const igraph_vector_t *weights,
                                   igraph_arpack_options_t *options) {
 
@@ -536,11 +593,19 @@ igraph_error_t igraph_eigenvector_centrality(const igraph_t *graph,
         options = igraph_arpack_options_get_default();
     }
 
-    if (directed && igraph_is_directed(graph)) {
-        return igraph_i_eigenvector_centrality_directed(graph, vector, value,
-                scale, weights, options);
-    } else {
+    if (mode != IGRAPH_ALL && mode != IGRAPH_OUT && mode != IGRAPH_IN) {
+        IGRAPH_ERROR("Invalid mode for eigenvector centrality.", IGRAPH_EINVAL);
+    }
+
+    if (! igraph_is_directed(graph)) {
+        mode = IGRAPH_ALL;
+    }
+
+    if (mode == IGRAPH_ALL) {
         return igraph_i_eigenvector_centrality_undirected(graph, vector, value,
-                scale, weights, options);
+                                                          weights, options);
+    } else {
+        return igraph_i_eigenvector_centrality_directed(graph, vector, value,
+                                                        mode, weights, options);
     }
 }