+++ tests/unit/igraph_eigenvector_centrality.c
@@ -17,19 +17,17 @@
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "igraph.h"
+#include <igraph.h>
 #include "test_utilities.h"
 
-#include <math.h>
-
-void print_and_destroy(igraph_t *g, igraph_vector_t *weights)
-{
+void print_and_destroy(igraph_t *g, igraph_neimode_t mode, igraph_vector_t *weights) {
     igraph_vector_t v;
     igraph_real_t value;
 
     igraph_vector_init(&v, 0);
-    igraph_eigenvector_centrality(g, &v, &value, /* directed */ true,
-                                  /* scale */ true, weights,
+    igraph_eigenvector_centrality(g, &v, &value,
+                                  mode,
+                                  weights,
                                   /* options */ NULL);
 
     printf("Eigenvalue: %g\n", value);
@@ -48,55 +46,98 @@ int main(void) {
 
     printf("Undirected graph with no vertices:\n");
     igraph_small(&g, 0, IGRAPH_UNDIRECTED, -1);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_ALL, NULL);
 
     printf("Directed graph with no vertices:\n");
     igraph_small(&g, 0, IGRAPH_DIRECTED, -1);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_OUT, NULL);
 
     printf("Undirected graph with no edges:\n");
     igraph_small(&g, 5, IGRAPH_UNDIRECTED, -1);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_ALL, NULL);
 
     printf("Directed graph with no edges:\n");
     igraph_small(&g, 5, IGRAPH_DIRECTED, -1);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_IN, NULL);
 
     printf("Undirected full graph:\n");
     igraph_full(&g, 5, IGRAPH_UNDIRECTED, /* loops */ false);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_ALL, NULL);
 
     printf("Directed full graph:\n");
     igraph_full(&g, 5, IGRAPH_DIRECTED, /* loops */ false);
-    print_and_destroy(&g, NULL);
+    print_and_destroy(&g, IGRAPH_OUT, NULL);
 
     printf("Undirected full graph with weights:\n");
     igraph_vector_init(&weights, 10);
     igraph_vector_fill(&weights, 1);
     igraph_full(&g, 5, IGRAPH_UNDIRECTED, /* loops */ false);
-    print_and_destroy(&g, &weights);
+    print_and_destroy(&g, IGRAPH_ALL, &weights);
     igraph_vector_destroy(&weights);
 
     printf("Directed full graph with weights:\n");
     igraph_full(&g, 5, IGRAPH_DIRECTED, /* loops */ false);
     igraph_vector_init(&weights, 20);
     igraph_vector_fill(&weights, 1);
-    print_and_destroy(&g, &weights);
+    print_and_destroy(&g, IGRAPH_OUT, &weights);
     igraph_vector_destroy(&weights);
 
     printf("Undirected star graph with weights:\n");
     igraph_vector_init(&weights, 4);
     igraph_vector_fill(&weights, 1);
     igraph_star(&g, 5, IGRAPH_STAR_UNDIRECTED, 0);
-    print_and_destroy(&g, &weights);
+    print_and_destroy(&g, IGRAPH_ALL, &weights);
     igraph_vector_destroy(&weights);
 
     printf("Directed star graph with weights:\n");
     igraph_star(&g, 5, IGRAPH_STAR_OUT, 0);
     igraph_vector_init(&weights, 4);
     igraph_vector_fill(&weights, 1);
-    print_and_destroy(&g, &weights);
+    print_and_destroy(&g, IGRAPH_OUT, &weights);
     igraph_vector_destroy(&weights);
+
+    printf("Small directed graph, OUT:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    print_and_destroy(&g, IGRAPH_OUT, NULL);
+
+    printf("Small directed graph, IN:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    print_and_destroy(&g, IGRAPH_IN, NULL);
+
+    printf("Small directed graph, ALL:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    print_and_destroy(&g, IGRAPH_ALL, NULL);
+
+    printf("Small directed graph with weights, OUT:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    igraph_vector_init_range(&weights, 1, igraph_ecount(&g)+1);
+    print_and_destroy(&g, IGRAPH_OUT, &weights);
+    igraph_vector_destroy(&weights);
+
+    printf("Small directed graph with weights, IN:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    igraph_vector_init_range(&weights, 1, igraph_ecount(&g)+1);
+    print_and_destroy(&g, IGRAPH_IN, &weights);
+    igraph_vector_destroy(&weights);
+
+    printf("Small directed graph with weights, ALL:\n");
+    igraph_small(&g, 4, IGRAPH_DIRECTED,
+                 0,1, 1,2, 2,3, 3,0, 1,3,
+                 -1);
+    igraph_vector_init_range(&weights, 1, igraph_ecount(&g)+1);
+    print_and_destroy(&g, IGRAPH_ALL, &weights);
+    igraph_vector_destroy(&weights);
+
 
     VERIFY_FINALLY_STACK();
 
@@ -104,11 +145,11 @@ int main(void) {
     igraph_vector_init(&weights, 2);
     igraph_vector_fill(&weights, 1);
     igraph_full(&g, 5, IGRAPH_DIRECTED, /* loops */ false);
-    CHECK_ERROR(igraph_eigenvector_centrality(&g, NULL, NULL, /* directed */ true,
-                                  /* scale */ true, &weights,
+    CHECK_ERROR(igraph_eigenvector_centrality(&g, NULL, NULL, /* mode */ IGRAPH_OUT,
+                                  &weights,
                                   /* options */ NULL), IGRAPH_EINVAL);
-    CHECK_ERROR(igraph_eigenvector_centrality(&g, NULL, NULL, /* directed */ false,
-                                  /* scale */ true, &weights,
+    CHECK_ERROR(igraph_eigenvector_centrality(&g, NULL, NULL, /* mode */ IGRAPH_ALL,
+                                  &weights,
                                   /* options */ NULL), IGRAPH_EINVAL);
     igraph_vector_destroy(&weights);
     igraph_destroy(&g);